add save state rgb #18 (-9 erase)

+++ color_module/color_module.h
@@ -4,7 +4,7 @@
 
 typedef struct 
 {
-    uint16_t h;
+    uint8_t h;
     uint8_t v;
     uint8_t s;
 }hsv_t;
